+++ Assets/Animations.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: ebdb237a55ebc1144a514c96dfe0af03
-NativeFormatImporter:
+guid: 3535772c57183fc4988f28bcf9cb69a5
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
