+++ Assets/Animations/Alien animation v2.anim
@@ -6,7 +6,7 @@ AnimationClip:
   m_CorrespondingSourceObject: {fileID: 0}
   m_PrefabInstance: {fileID: 0}
   m_PrefabAsset: {fileID: 0}
-  m_Name: Alien animation
+  m_Name: Alien animation v2
   serializedVersion: 6
   m_Legacy: 0
   m_Compressed: 0
