+++ Assets/Animations/Alien animation v2.anim.meta
@@ -1,7 +1,8 @@
 fileFormatVersion: 2
-guid: f4091d4756b77114d9b5592907af8da9
-PrefabImporter:
+guid: 250e2f9b02c84b74fa93ced0264423e4
+NativeFormatImporter:
   externalObjects: {}
+  mainObjectFileID: 7400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
