+++ Assets/Animations/spritesheet-v2_0 1.controller.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 44caa3fee6dec0f46a36f2cb47d80351
+guid: fbf6e1f9eff50b642bc6056fb4bce6a2
 NativeFormatImporter:
   externalObjects: {}
-  mainObjectFileID: 7400000
+  mainObjectFileID: 9100000
   userData: 
   assetBundleName: 
   assetBundleVariant:
