--- Assets/Prefabs/Alien.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 34d3ce00c0dc13743b15715ef6ccdc03
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
