--- Assets/Prefabs/ShipBullet.prefab.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: ec1c6e8f113fc5b4e9cc5d4b2ccc694e
-PrefabImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
