+++ Assets/Sprites/bullet.png.meta
