--- codice nemici e bullet/EnemyScript.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bf4d37c50dcc39e4ba5b0c0df62403db
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
