--- prefab nemici/prefab alien/New Animation alien.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0cb90734138a03b48af6758607cd53c5
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
